requirements.txt
Django>=1.8,<1.9
djangorestframework==3.1.3
psycopg2
python3-memcached==1.51
raven==5.0.0
bootstrap-admin==0.3.3
django-gravatar2==1.1.3
django-pgjson
pillow
six
aiohttp
oauthlib==0.7.2
requests
beautifulsoup4
tqdm
djorm-pgarray